fixed INtegrationTests

+++ KooliProjekt.IntegrationTests/Helpers/TestBase.cs
@@ -1,7 +1,8 @@
-﻿using System;
-using Kooliprojekt.Data;
-using KooliProjekt.Data;
+﻿using KooliProjekt.Data;
 using Microsoft.AspNetCore.Mvc.Testing;
+using System;
+using System.Net.Http;
+using Kooliprojekt.Data;
 
 namespace KooliProjekt.IntegrationTests.Helpers
 {
@@ -14,12 +15,24 @@ namespace KooliProjekt.IntegrationTests.Helpers
             Factory = new TestApplicationFactory<FakeStartup>();
         }
 
-        public void Dispose()
+        public HttpClient CreateClient()
         {
-            var dbContext = (ApplicationDbContext)Factory.Services.GetService(typeof(ApplicationDbContext));
-            dbContext.Database.EnsureDeleted();
+            return Factory.CreateClient();
         }
 
-        // Add your other helper methods here
+        public ApplicationDbContext GetDbContext()
+        {
+            return (ApplicationDbContext)Factory.Services.GetService(typeof(ApplicationDbContext));
+        }
+
+        public void Dispose()
+        {
+            var dbContext = GetDbContext();
+            if (dbContext != null)
+            {
+                dbContext.Database.EnsureDeleted();
+            }
+            Factory.Dispose();
+        }
     }
 }
